feat(web): refactor code #21
